Refactored and added new docs for `Program`
